requirements.txt
pytz
flask
EdgeGPT
waitress
nest-asyncio
idna==3.4
bs4==0.0.1
lxml==4.9.1
poetry==1.2.2
requests==2.28.1
urllib3==1.26.13
tls-client==0.1.5
soupsieve==2.3.2.post1
beautifulsoup4==4.11.1
charset-normalizer==2.1.1
python-telegram-bot==20.2